+++ src/records/WebmasterSettingsRecord.php
@@ -11,6 +11,10 @@ use craft\db\ActiveRecord;
  */
 class WebmasterSettingsRecord extends ActiveRecord
 {
+    use InvalidatesCacheTagTrait;
+
+    public const CACHE_TAG = 'beacon_site_settings';
+
     public static function tableName(): string
     {
         return '{{%beacon_webmaster_settings}}';